Avoid use of size_t where a 64 bit value is specifically required
size_t is appropriate for sizes of objects/indexes, but a fixed width
type like uint64_t should be used instead if it is _necessary_ that the
value be 64 bits, even e.g., on a 32 bit host.

+++ src/pdu/display/bitwidth_histogram.cc
@@ -35,8 +35,8 @@ BitWidthHistogram& BitWidthHistogram::operator+=(
 }
 
 void BitWidthHistogram::print(bool percent, bool human) const {
-    size_t totalCount = 0;
-    size_t totalSize = 0;
+    uint64_t totalCount = 0;
+    uint64_t totalSize = 0;
     int index = 0;
     for (auto count : values) {
         totalCount += count;
